refactor: 테스트 패키지 변경 및 필드 이름 변경
