+++ src/main/java/db/DataBase.java
@@ -1,14 +1,13 @@
 package db;
 
+import com.google.common.collect.Maps;
+import model.User;
+
 import java.util.Collection;
 import java.util.Map;
 
-import com.google.common.collect.Maps;
-
-import model.User;
-
 public class DataBase {
-    private static Map<String, User> users = Maps.newHashMap();
+    private static final Map<String, User> users = Maps.newHashMap();
 
     public static void addUser(User user) {
         users.put(user.getUserId(), user);
