+++ src/test/java/web/request/HttpRequestTest.java
@@ -1,4 +1,4 @@
-package web;
+package web.request;
 
 import exception.InvalidHttpRequestException;
 import org.assertj.core.api.Assertions;
@@ -7,7 +7,6 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import web.request.HttpRequest;
 
 import java.io.*;
