+++ src/test/java/web/request/RequestLineTest.java
@@ -1,4 +1,4 @@
-package web;
+package web.request;
 
 import exception.InvalidRequestLineException;
 import org.assertj.core.api.Assertions;
@@ -6,7 +6,6 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.NullAndEmptySource;
-import web.request.RequestLine;
 
 public class RequestLineTest {
     @Test
