+++ src/test/java/web/request/RequestPathTest.java
@@ -1,4 +1,4 @@
-package web;
+package web.request;
 
 import exception.InvalidRequestParamsException;
 import exception.InvalidRequestPathException;
@@ -7,7 +7,6 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.NullAndEmptySource;
-import web.request.RequestPath;
 
 public class RequestPathTest {
     @Test
